hnsw ops doc
